+++ RELEASENOTES.md
@@ -124,6 +124,8 @@
     *   MP3: Fix bitrate reporting for files with Xing and VBRI header.
     *   Matroska: Handle tracks defined in the last cluster
         ([#3250](https://github.com/androidx/media/issues/3250)).
+    *   MP4: Allow variable length `sgpd` boxes in fragmented MP4 files
+        ([#3243](https://github.com/androidx/media/pull/3243)).
 *   Inspector:
 *   Audio:
     *   Add a 100ms grace period in ExoPlayer's audio renderers when